requirements-build.txt
srsly==2.4.6
tqdm==4.65.0
arxiv==1.4.2
httpx==0.23.3
spacy==3.5.0
retry==0.9.2
wandb==0.14.2
embetter==0.5.0
sentence-transformers==2.2.2
simsity==0.5.5
questionary==1.10.0
radicli==0.0.25
lazylines==0.0.4
rich==13.4.2
lunr==0.6.2
skops==0.6.0
python-dotenv==1.0.0
scikit-learn==1.3
transformers==4.32.1
torch==2.1.2
openai==1.9.0